Add fuzzer for string length invariant
