+++ fuzz/fuzz_targets/main.rs
@@ -8,37 +8,12 @@ use libfuzzer_sys::fuzz_target;
 use rcl::error::Error;
 use rcl::error::Result;
 use rcl::eval::Evaluator;
-use rcl::loader::{Document, Filesystem, Loader, PathLookup};
+use rcl::loader::{Document, Filesystem, Loader, PathLookup, VoidFilesystem};
 use rcl::markup::MarkupMode;
 use rcl::pprint;
 use rcl::runtime::Value;
 use rcl::source::{Inputs, Span};
-use rcl::tracer::Tracer;
-
-/// Tracer that ignores its messages.
-pub struct VoidTracer;
-
-impl Tracer for VoidTracer {
-    fn trace(&mut self, _inputs: &Inputs, _span: Span, _message: Rc<Value>) {}
-}
-
-/// Filesystem that fails to load anything.
-///
-/// TODO: We could populate files from fuzz inputs to test imports.
-struct VoidFilesystem;
-
-impl Filesystem for VoidFilesystem {
-    fn resolve(&self, _: &str, _: &str) -> Result<PathLookup> {
-        Error::new("Void filesystem does not load files.").err()
-    }
-    fn resolve_entrypoint(&self, _: &str) -> Result<PathLookup> {
-        Error::new("Void filesystem does not load files.").err()
-    }
-    fn load(&self, _: PathLookup) -> Result<Document> {
-        Error::new("Void filesystem does not load files.").err()
-    }
-
-}
+use rcl::tracer::VoidTracer;
 
 #[derive(Debug)]
 enum Mode {
